+++ java/rocksjni/statisticsjni.h
@@ -12,12 +12,14 @@
 #include <memory>
 #include <set>
 #include <string>
-#include "rocksdb/statistics.h"
+
 #include "monitoring/statistics.h"
+#include "monitoring/statistics_impl.h"
+#include "rocksdb/statistics.h"
 
 namespace ROCKSDB_NAMESPACE {
 
-class StatisticsJni : public StatisticsImpl {
+class StatisticsJni : public StatisticsImpl<> {
  public:
   StatisticsJni(std::shared_ptr<Statistics> stats);
   StatisticsJni(std::shared_ptr<Statistics> stats,
@@ -26,8 +28,8 @@ class StatisticsJni : public StatisticsImpl {
 
  private:
   const std::set<uint32_t> m_ignore_histograms;
- };
+};
 
- }  // namespace ROCKSDB_NAMESPACE
+}  // namespace ROCKSDB_NAMESPACE
 
 #endif  // JAVA_ROCKSJNI_STATISTICSJNI_H_
