+++ tools/block_cache_analyzer/block_cache_trace_analyzer.cc
@@ -22,6 +22,7 @@
 #include "rocksdb/system_clock.h"
 #include "rocksdb/trace_record.h"
 #include "util/gflags_compat.h"
+#include "util/random.h"
 #include "util/string_util.h"
 
 using GFLAGS_NAMESPACE::ParseCommandLineFlags;
